Set transient duration shorter than its secret's lease duration.
If an async refresh fails, it is possible that `get_secret()` could return
`null` between its value expiring and a new secret being refreshed.

In conjunction with the async refresh process, setting the transient
duration's slightly less than the secret's lease duration should remove this
possibility in normal operation.

If a failure occurs after all of this, then there's a bigger problem with the
connection that will need manual intervention to resolve.